// design-code-dashboard/src/app/core/models/design.models.ts
export interface SelectOption {
  label: string;
  value: string;
}